Bit of a hack
