+++ tests/test_simple.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import bream
+from bream.core import EncodeError
 
 
 def test_scalar_encode() -> None:
@@ -19,28 +20,11 @@ def test_list_encode() -> None:
     assert x_encoded is not x
 
 
-def test_dict_encode() -> None:
-    fmt = bream.SerialisationFormat(codecs=())
-    x = {"a": None, "b": 2, "c": 4.2, "d": "moo"}
-    x_encoded = bream.encode(x, fmt)
-    assert x_encoded == x
-    assert x_encoded is not x
-
-
-def test_nested_structure_encode() -> None:
-    fmt = bream.SerialisationFormat(codecs=())
-    x = [None, 2, 4.2, "moo"]
-    y = {"a": None, "b": 2, "c": 4.2, "d": "moo", "e": x}
-    y_encoded = bream.encode(y, fmt)
-    assert y_encoded == y
-    assert y_encoded is not y
-
-
 def test_unsupported_encode() -> None:
     fmt = bream.SerialisationFormat(codecs=())
-    assert bream.encode(1j, fmt) == bream.NoEncoderAvailable(value=1j)
-    assert bream.encode((42,), fmt) == bream.NoEncoderAvailable(value=(42,))
-    assert bream.encode({3}, fmt) == bream.NoEncoderAvailable(value={3})
+    assert bream.encode(1j, fmt) == bream.core.NoEncoderAvailable(value=1j)
+    assert bream.encode((42,), fmt) == bream.core.NoEncoderAvailable(value=(42,))
+    assert bream.encode({3}, fmt) == bream.core.NoEncoderAvailable(value={3})
 
 
 def test_scalar_decode() -> None:
@@ -61,31 +45,25 @@ def test_list_decode() -> None:
     assert x_decoded is not x
 
 
-def test_dict_decode() -> None:
+def test_nested_structure_round_trip() -> None:
     fmt = bream.SerialisationFormat(codecs=())
-    bream_spec = 0
-    x: dict[str, bream.JsonType] = {"a": None, "b": 2, "c": 4.2, "d": "moo"}
-    x_decoded = bream.decode(x, fmt, bream_spec)
-    assert x_decoded == x
-    assert x_decoded is not x
-
-
-def test_nested_structure_decode() -> None:
-    fmt = bream.SerialisationFormat(codecs=())
-    bream_spec = 0
-    x: list[bream.JsonType] = [None, 2, 4.2, "moo"]
-    y: dict[str, bream.JsonType] = {"a": None, "b": 2, "c": 4.2, "d": "moo", "e": x}
-    y_decoded = bream.decode(y, fmt, bream_spec)
-    assert y_decoded == y
+    x = [None, 2, 4.2, "moo"]
+    y = ["a", None, x]
+    y_encoded = bream.encode(y, fmt)
+    assert not isinstance(y_encoded, EncodeError)
+    assert y_encoded == y
+    assert y_encoded is not y
+    y_decoded = bream.decode(y_encoded, fmt, bream.core.BREAM_SPEC)
     assert y_decoded is not y
+    assert y_decoded == y
 
 
 def test_simple_document_round_trip() -> None:
     fmt = bream.SerialisationFormat(codecs=())
-    for x in (None, 2, 4.2, "moo", [1, None, 3], {"a": [3, 4], "b": {"c": 4.2}}):
+    for x in (None, 2, 4.2, "moo", [1, None, 3], ["a", [3, 4], "b", ["c", 4.2]]):
         document = bream.encode_to_document(x, fmt)
         assert not isinstance(document, bream.EncodeError)
-        assert document == {"_bream_spec": bream.BREAM_SPEC, "_payload": x}
+        assert document == {"_bream_spec": bream.core.BREAM_SPEC, "_payload": x}
         new_x = bream.decode_document(document, fmt)
         assert new_x == x
         if isinstance(x, list | dict):
